requirements.txt
numpy
opencv-python
pillow
imutils